% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ais}
\alias{ais}
\title{AIS data, 1 July to 31 October, 2018}
\format{
An object of class \code{data.frame} with 259727 rows and 14 columns.
}
\usage{
ais
}
\description{
AIS data, 1 July to 31 October, 2018
}
\keyword{datasets}
