% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision_curve.R
\name{collision_curve}
\alias{collision_curve}
\title{Produce a P(Collision) ~ Speed curve}
\usage{
collision_curve(
  speeds = seq(0, 30, length = 1000),
  c1 = -0.2002002,
  c2 = 11.80241,
  asymptote = 0.9
)
}
\arguments{
\item{speeds}{A vector of speeds at which to predict P(Lethality).}

\item{c1}{First parameter.}

\item{c2}{Second parameter.}

\item{asymptote}{Upper asymptote of the function.}
}
\value{
A \code{data.frame}.
}
\description{
This function returns the Probability of Collision for a range of speeds based on
a logistic function of form: \code{P(Collision) = asymptote/ (1 + exp(c1*(speeds - c2)))}.
The defaults are the values used in \emph{Keen et al. (2023)} for large ships > 180m.
}
