% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_results.R
\name{compile_results}
\alias{compile_results}
\title{Compile results from outcomes function}
\usage{
compile_results(outcome_dir)
}
\arguments{
\item{outcome_dir}{dir}
}
\value{
A big list.
}
\description{
Compile results from outcomes function
}
