% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Draw an ellipse}
\usage{
ellipse(x, y, width, height = width, theta = 0, npoints = 10, plot = FALSE)
}
\arguments{
\item{x}{desc}

\item{y}{desc}

\item{width}{desc}

\item{height}{desc}

\item{theta}{desc}

\item{npoints}{desc}

\item{plot}{desc}
}
\value{
desc
}
\description{
Draw an ellipse
}
