% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_diagnostics.R
\name{encounter_diagnostics}
\alias{encounter_diagnostics}
\title{Plot diagnostics for a single encounter-rate transit}
\usage{
encounter_diagnostics(transits, id = NULL, wait = TRUE)
}
\arguments{
\item{transits}{desc}

\item{id}{desc}

\item{wait}{desc}
}
\value{
desc
}
\description{
Plot diagnostics for a single encounter-rate transit
}
