% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_rate.R
\name{encounter_rate}
\alias{encounter_rate}
\title{New encounter rate function}
\usage{
encounter_rate(
  vessels,
  whales,
  outcome_dir,
  month_batches = list(winter = c(0:4, 11:12), summer = 5:10),
  speed_restriction = NULL,
  lengths_restricted = 0,
  new_speeds = NULL,
  runs = 100,
  iterations = 100,
  toplot = TRUE
)
}
\arguments{
\item{vessels}{desc}

\item{whales}{desc}

\item{outcome_dir}{desc}

\item{month_batches}{desc}

\item{speed_restriction}{Apply a maximum speed, converting all vessel positions
that exceed that maximum to the maximum.}

\item{lengths_restricted}{Minimum length affected by speed restriction, if any.}

\item{new_speeds}{desc}

\item{runs}{desc}

\item{iterations}{desc}
}
\value{
desc (note that if runs == 1, a detailed dataframe will be returned (about 2.5 million rows))
}
\description{
New encounter rate function
}
