% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_wrapper.R
\name{encounter_wrapper}
\alias{encounter_wrapper}
\title{Wrapper for encounter simulator}
\usage{
encounter_wrapper(
  vessel_source,
  whale_source,
  outcome_dir,
  process_months = TRUE,
  runs = 100,
  iterations = 100
)
}
\arguments{
\item{vessel_source}{desc}

\item{whale_source}{desc}

\item{outcome_dir}{desc}

\item{process_months}{desc}

\item{runs}{desc}

\item{iterations}{desc}
}
\value{
desc
}
\description{
Wrapper for encounter simulator
}
