% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fin_params.R
\name{fin_params}
\alias{fin_params}
\title{Fin whale parameters}
\usage{
fin_params(
  length_min = 10,
  length_max = 26,
  length_mean = 20,
  length_sd = 1.65,
  width_factor = 0.2074,
  speed_day_min = 0.27,
  speed_day_max = 2.22,
  speed_day_mean = 1.140301,
  speed_day_sd = 0.4375298,
  speed_night_min = 0.55,
  speed_night_max = 3.04,
  speed_night_mean = 1.809226,
  speed_night_sd = 0.5718535,
  delta_day_mean = 0.8087532,
  delta_day_sd = 1.332026,
  delta_night_mean = 1.494988,
  delta_night_sd = 1.386166
)
}
\arguments{
\item{length_min}{Whale dimensions will be drawn from
a truncated normal distribution (source: Keen et al 2021)}

\item{length_max}{desc}

\item{length_mean}{desc}

\item{length_sd}{desc}

\item{width_factor}{will be used to scale length to find width}

\item{speed_day_min}{Whale speed will be drawn from
a truncated normal distribution (source: Hendricks et al 2021)}

\item{speed_day_max}{desc}

\item{speed_day_mean}{desc}

\item{speed_day_sd}{desc}

\item{speed_night_min}{desc}

\item{speed_night_max}{desc}

\item{speed_night_mean}{desc}

\item{speed_night_sd}{desc}

\item{delta_day_mean}{Whale track variability defined as the
standard deviation of changes in whale course every 60 seconds;
values will be drawn from a normal distribution then reverse-log-transformed
(source: Hendricks et al. 2021)}

\item{delta_day_sd}{desc}

\item{delta_night_mean}{desc}

\item{delta_night_sd}{desc}
}
\value{

}
\description{
Fin whale parameters
}
