% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fin_whale_habitat}
\alias{fin_whale_habitat}
\title{Coordinates for the polygon representing prime fin whale habitat
(representing the waterway with the highest density of fin whales, as found by Keen et al. 2021)}
\format{
An object of class \code{data.frame} with 60 rows and 2 columns.
}
\usage{
fin_whale_habitat
}
\description{
Coordinates for the polygon representing prime fin whale habitat
(representing the waterway with the highest density of fin whales, as found by Keen et al. 2021)
}
\keyword{datasets}
