% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_outcomes.R
\name{gather_outcomes}
\alias{gather_outcomes}
\title{Gather and concatentate outcome \code{RData} files}
\usage{
gather_outcomes(outcome_dir, verbose = TRUE)
}
\arguments{
\item{outcome_dir}{desc}

\item{verbose}{desc}
}
\value{

}
\description{
Gather and concatentate outcome \code{RData} files
}
