% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seafloor.R
\name{get_seafloor}
\alias{get_seafloor}
\title{Get seafloor features at a spatial point}
\usage{
get_seafloor(lon, lat, seafloor)
}
\arguments{
\item{lon}{desc}

\item{lat}{desc}

\item{seafloor}{desc}
}
\value{

}
\description{
Get seafloor features at a spatial point
}
