% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grid_kfs}
\alias{grid_kfs}
\title{1-km2 spatial grid for the study area (Gitga'at waters, Kitimat Fjord System, BC, Canada)}
\format{
An object of class \code{data.frame} with 1409 rows and 14 columns.
}
\usage{
grid_kfs
}
\description{
1-km2 spatial grid for the study area (Gitga'at waters, Kitimat Fjord System, BC, Canada)
}
\keyword{datasets}
