% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humpback_params.R
\name{humpback_params}
\alias{humpback_params}
\title{Humpback whale parameters}
\usage{
humpback_params(
  length_min = 8.5344,
  length_max = 15.448,
  length_mean = 12.1158,
  length_sd = 1.524,
  width_factor = 0.23778,
  speed_day_min = 0.13888,
  speed_day_max = 1.166,
  speed_day_mean = 0.7111,
  speed_day_sd = 0.20833,
  speed_night_min = 0.22035,
  speed_night_max = 1.85,
  speed_night_mean = 1.128,
  speed_night_sd = 0.2247,
  delta_day_mean = 0.8087532,
  delta_day_sd = 1.332026,
  delta_night_mean = 1.494988,
  delta_night_sd = 1.386166
)
}
\arguments{
\item{length_min}{Whale dimensions will be drawn from
a truncated normal distribution (source: TBD)}

\item{length_max}{desc}

\item{length_mean}{desc}

\item{length_sd}{desc}

\item{width_factor}{will be used to scale length to find width}

\item{speed_day_min}{Whale speed will be drawn from
a truncated normal distribution (source: TBD)}

\item{speed_day_max}{desc}

\item{speed_day_mean}{desc}

\item{speed_day_sd}{desc}

\item{speed_night_min}{desc}

\item{speed_night_max}{desc}

\item{speed_night_mean}{desc}

\item{speed_night_sd}{desc}

\item{delta_day_mean}{Whale track variability defined as the
standard deviation of changes in whale course every 60 seconds;
values will be drawn from a normal distribution then reverse-log-transformed
(source: using fin whale data -- Hendricks et al. 2021)}

\item{delta_day_sd}{desc}

\item{delta_night_mean}{desc}

\item{delta_night_sd}{desc}
}
\value{

}
\description{
Humpback whale parameters
}
