% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lethality_curve.R
\name{lethality_curve}
\alias{lethality_curve}
\title{Produce a P(Lethality) ~ Speed curve}
\usage{
lethality_curve(
  speeds = seq(0, 30, length = 1000),
  c1 = -1.241241,
  c2 = 0.2712432,
  asymptote = 1
)
}
\arguments{
\item{speeds}{A vector of speeds at which to predict P(Lethality).}

\item{c1}{First parameter}

\item{c2}{Second parameter}

\item{asymptote}{Upper asymptote of the function.}
}
\value{
A \code{data.frame}.
}
\description{
This function returns the Probability of Lethality for a range of speeds based on
a logistic function of form: \code{P(Lethality) = asymptote/ (1 + exp(-1*(c1 + (c2*speeds))))}.
The defaults are the values used in \emph{Keen et al. (2023)} for large ships > 180m.
}
