% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Create a spatial grid}
\usage{
make_grid(xlims, ylims, grid_int)
}
\arguments{
\item{xlims}{desc}

\item{ylims}{desc}

\item{grid_int}{adjust grid_int (km) to get mean grid km2 of 1.00 km2}
}
\value{

}
\description{
Create a spatial grid
}
