% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitigate.R
\name{mitigate}
\alias{mitigate}
\title{Apply mitigation filters to a list of outcomes}
\usage{
mitigate(outcomes, mitigation_filters)
}
\arguments{
\item{outcomes}{desc}

\item{mitigation_filters}{desc}
}
\value{
desc
}
\description{
Apply mitigation filters to a list of outcomes
}
