% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitigate_loop.R
\name{mitigate_loop}
\alias{mitigate_loop}
\title{Loop through months to find most efficacious mitigation window}
\usage{
mitigate_loop(
  outcomes,
  mitigation_duration = 1,
  reschedule = FALSE,
  vessels = c("Tanker > 100m", "Cargo > 100m", "Passenger > 100m",
    "LNG Canada tanker in-heel", "LNG Canada tanker in-product",
    "LNG Canada tug in-heel", "LNG Canada tug in-product", "Cedar LNG tanker in-heel",
    "Cedar LNG tanker in-product", "Cedar LNG tug in-heel", "Cedar LNG tug in-product")
)
}
\arguments{
\item{outcomes}{The result of \code{outcome_predict()}.}

\item{mitigation_duration}{Number indicating the number of months to use as a mitigation period.
Keen et al. (2023) tests durations of 1, 2, and 3 months.}

\item{reschedule}{Boolean, with default \code{FALSE}, indicating whether traffic during the mitigation window
should be rescheduled (\code{reschedule == TRUE}) or cancelled (\verb{FALSE)}.}

\item{vessels}{A character vector indicating for which vessel types this mitigation action applies.}
}
\value{
See the \code{shipstrike} package vignette & Keen et al. (2023) for details.
}
\description{
To test the effects of seasonally displacing ship traffic
(i.e., rescheduling transits from a given month into other months)
\emph{or} seasonally cancelling ship traffic
(i.e., a moratorium), you can use the \code{shipstrike} function \code{mitigate_loop()},
which applies the displacement window to each candidate month in a loop.
This allows us to determine which month would be the most efficacious target for mitigation.
}
