% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{outcome}
\alias{outcome}
\title{Core function for whale-ship collision outcome}
\usage{
outcome(p_whale, p_encounter, p_surface, p_avoidance, p_lethality)
}
\arguments{
\item{p_whale}{desc}

\item{p_encounter}{desc}

\item{p_surface}{desc}

\item{p_avoidance}{desc}

\item{p_lethality}{desc}
}
\value{

}
\description{
Core function for whale-ship collision outcome
}
