% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_histograms.R
\name{outcome_histograms}
\alias{outcome_histograms}
\title{Create histograms summarizing whale-ship outcomes}
\usage{
outcome_histograms(
  outcomes,
  surface_var = "surface2",
  collision_var = "collision2.2",
  mortality_var = "mortality2.2",
  plot_all = FALSE,
  orientation = "vertical"
)
}
\arguments{
\item{outcomes}{A \code{data.frame} of results from the \code{shipstrike::gather_outcomes()} function.}

\item{surface_var}{desc}

\item{collision_var}{desc}

\item{mortality_var}{desc}

\item{orientation}{desc}
}
\value{
desc
}
\description{
Create histograms summarizing whale-ship outcomes
}
