% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_map.R
\name{outcome_map}
\alias{outcome_map}
\title{Plot map of outcomes}
\usage{
outcome_map(
  outcomes_grid,
  base_map = NULL,
  varplot = "cooccurrence",
  varfacet = NULL,
  spp = NULL,
  years = NULL,
  channels = NULL,
  vessels = NULL,
  months = NULL,
  diels = NULL,
  col_min = "white",
  col_max = "darkorange4"
)
}
\arguments{
\item{outcomes_grid}{The \code{outcomes_grid} output from \code{outcome_predict()}.}

\item{base_map}{Base map, provided as a \code{ggplot2} object. The default is to
use \code{bangarang::gg_kfs()} of the Kitimat Fjord System (BC, Canada).}

\item{varplot}{The variable to plot. See output of \code{outcome_predict()}.}

\item{varfacet}{The variable by which to facet, if any.}

\item{spp}{Species to filter by.}

\item{years}{Year to filter by.}

\item{channels}{\code{channel} to filter by.}

\item{vessels}{\code{vessels} to filter by.}

\item{months}{\code{months} (numerics) to filter by.}

\item{diels}{Diel period \code{"night"} or \code{"day"}).}

\item{col_min}{Specify minimum color of color gradient scale.}

\item{col_max}{Specify maximum color of color gradient scale.}
}
\value{
A \code{ggplot} object.
}
\description{
This function produces a map of a risk metric.
The default is to show cooccurrences for all month-diel-vessels.
}
