% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_melt.R
\name{outcome_melt}
\alias{outcome_melt}
\title{Melt the result of \code{predict_outcomes()} to use with \code{tidyverse} verbs.}
\usage{
outcome_melt(
  mr,
  events = c("cooccurrence", "encounter", "surface", "surface2", "collision1.1",
    "collision1.2", "collision1.3", "collision1.4", "collision2.1", "collision2.2",
    "collision2.3", "collision2.4", "mortality1.1", "mortality1.2", "mortality1.3",
    "mortality1.4", "mortality2.1", "mortality2.2", "mortality2.3", "mortality2.4")
)
}
\arguments{
\item{mr}{A \code{data.frame} of results from the \code{shipstrike::gather_outcomes()} function.}

\item{events}{A character vector of outcome types to retain in the melted dataset.}
}
\value{
A melted \code{data.frame}
}
\description{
This function may be handy if you want to conduct your own tailored analyses.
It melts the results data into a tidyverse-friendly format in which every row
is a single prediction. Every function above calls this function internally.
}
