% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_predict.R
\name{outcome_predict}
\alias{outcome_predict}
\title{Predict whale-ship interaction outcomes}
\usage{
outcome_predict(
  traffic,
  scale_factors = NULL,
  whale,
  seasonal = NULL,
  p_encounter_dir = NULL,
  surface,
  avoidance,
  lethality,
  outcome_dir,
  asymptote_scaling = NULL,
  month_batches = list(winter = c(0:4, 11:12), summer = 5:10),
  species = "fw",
  year = 2019,
  iterations = 1000
)
}
\arguments{
\item{traffic}{desc}

\item{scale_factors}{desc}

\item{whale}{desc}

\item{seasonal}{desc}

\item{p_encounter_dir}{desc}

\item{surface}{desc}

\item{avoidance}{desc}

\item{lethality}{desc}

\item{outcome_dir}{desc}

\item{asymptote_scaling}{desc}

\item{species}{desc}

\item{year}{desc}

\item{iterations}{desc}
}
\value{
.RData files saved to outcome_dir
}
\description{
Predict whale-ship interaction outcomes
}
