% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_shares.R
\name{outcome_shares}
\alias{outcome_shares}
\title{Share of risk by vessel type, channel, month, and diel period}
\usage{
outcome_shares(outcomes, verbose = TRUE)
}
\arguments{
\item{outcomes}{desc}
}
\value{
desc
}
\description{
Share of risk by vessel type, channel, month, and diel period
}
