% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_validation.R
\name{outcome_validation}
\alias{outcome_validation}
\title{Outcome validation simulation}
\usage{
outcome_validation(
  outcomes,
  eventcol = "mortality2.2",
  observations = 0,
  scaling_factor = 1,
  years = 10,
  months = 6:9,
  iterations = 1000,
  toplot = FALSE
)
}
\arguments{
\item{outcomes}{A \code{data.frame} of results from the \code{shipstrike::gather_outcomes()} function.}

\item{observations}{The number of observed strike events during observation period.}

\item{scaling_factor}{Deprecated -- keep at 1.}

\item{years}{Years of observation.}

\item{months}{Months of effort during each year.}

\item{iterations}{Number of iterations to use for randomization routine.}

\item{toplot}{Produce plots?}
}
\value{
A named \code{list} including \code{ggplot2} objects, as well as output printed to console.
}
\description{
This function replicates the validation exercise from Keen et al. (2023)
to determine what our strike detection rate would need to be in order for
our results to be plausible given that we have never observed a strike in our study area.
}
