% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbr.R
\name{pbr}
\alias{pbr}
\title{Estimate Potential Biological Removal (PBR)}
\usage{
pbr(N, CV, Rmax = 0.08, Fr = 0.5, toplot = FALSE)
}
\arguments{
\item{N}{Abundance estimate}

\item{CV}{CV of abundance estimate}

\item{Rmax}{Maximum rate of increase. Default is that for cetaceans recommended by NMFS (2016).}

\item{Fr}{Recovery factor. Default is that for a depleted cetacean stock, recommended by NMFS (2016).}

\item{toplot}{Boolean, whether or not to plot the distribution of the abundance estimate.}
}
\value{
A simple named \code{list} with slots \verb{$PBR}, \verb{$Nmin}, \verb{$Rmax}, \verb{$Fr}, and \verb{$Nmedian}.
}
\description{
Assuming a log-normal distribution of abundance estimates.
}
