% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcomes_channels.R
\name{plot_outcomes_channels}
\alias{plot_outcomes_channels}
\title{Plot raw outcomes: channels}
\usage{
plot_outcomes_channels(outcomes)
}
\arguments{
\item{outcomes}{desc}
}
\value{
desc
}
\description{
Plot raw outcomes: channels
}
