% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcomes_diels.R
\name{plot_outcomes_diels}
\alias{plot_outcomes_diels}
\title{Plot raw outcomes: diels}
\usage{
plot_outcomes_diels(outcomes)
}
\arguments{
\item{outcomes}{desc}
}
\value{
desc
}
\description{
Plot raw outcomes: diels
}
