% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcomes_months.R
\name{plot_outcomes_months}
\alias{plot_outcomes_months}
\title{Plot raw outcomes: monthly}
\usage{
plot_outcomes_months(outcomes)
}
\arguments{
\item{outcomes}{desc}
}
\value{
desc
}
\description{
Plot raw outcomes: monthly
}
