% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_outcomes_vessels.R
\name{plot_outcomes_vessels}
\alias{plot_outcomes_vessels}
\title{Plot raw outcomes: vessels}
\usage{
plot_outcomes_vessels(outcomes)
}
\arguments{
\item{outcomes}{desc}
}
\value{
desc
}
\description{
Plot raw outcomes: vessels
}
