% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_outcomes.R
\name{predict_outcomes}
\alias{predict_outcomes}
\title{Predict whale-ship interaction outcomes}
\usage{
predict_outcomes(
  traffic,
  whale,
  seasonal,
  p_encounter_dir,
  surface,
  avoidance,
  lethality,
  outcome_dir,
  iterations = 1000
)
}
\arguments{
\item{traffic}{desc}

\item{whale}{desc}

\item{seasonal}{desc}

\item{p_encounter_dir}{desc}

\item{surface}{desc}

\item{avoidance}{desc}

\item{lethality}{desc}

\item{outcome_dir}{desc}

\item{iterations}{desc}
}
\value{
.RData files saved to outcome_dir
}
\description{
Predict whale-ship interaction outcomes
}
