% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_vessel.R
\name{simulate_vessel}
\alias{simulate_vessel}
\title{Simulate a set of vessel transits on a spatial grid}
\usage{
simulate_vessel(grids, vessels, toplot = TRUE, verbose = TRUE)
}
\arguments{
\item{grids}{A spatial grid of cells, as produced by \code{shipstrike::make_grid()} (see documentation)
(or at least matching the format thereof).}

\item{vessels}{A list of lists, each one representing settings for the simulation of a set of vessel transits.
The expected format of each list is as follows:
\itemize{
\item \code{type} A chatacter string name for this type of vessel, whatever you want: e.g., \code{"fishing"}, \code{"tanker"}, etc.
\item \code{n} A numeric indicating the number of transits to simulate.
\item \code{size_min} Minimum vessel size (meters)
\item \code{size_max} Maximum vessel size (meters); for each transit, a size will be
drawn from a uniform-random distribution between the min and the max.
\item \code{speed_min} Minimum vessel speed (knots)
\item \code{speed_max} Maximum vessel speed (knots); as with size, the speed of each transit will
be drawn from a uniform random distribution between the min and the max.
\item \code{width} A scaling factor, between 0 and 1, used to scale the length of the vessel
to get its beam width. Typical values for a large ship range from 0.05 to 0.15.
\item \code{draft} A scaling factor, between 0 and 1, used to scale the length of the vessel
to get its draft. Typical values are somewhere between 0.03 and 0.10.
\item \code{hour} The hour of day on which to begin the transit; if not provided or \code{NULL},
a random start hour will be chosen.
\item \code{months} The months of year in which to simualte transits; if nothing is provided, or if \code{NULL},
the month for each transit will be randomly chosen.
This can be handy if you want to simulate the effects of adding vessel traffic at only certain times of year.
Regardless of whether \code{month} is specified, the day of month for each transit is then randomly chosen.
\item \code{direction} A means of alternating or reversing the direction transited along the vessel's course; ignored if
\code{course} below is \code{NULL}. If \code{direction} is \code{0} or \code{NULL} or missing,
the vessel always travels along the course in the sequence indicated
within the \code{course} \code{data.frame}, starting at the first pair of coordinates.
If \code{direction} is \code{1}, the course will be reversed.
If \code{direction} is \code{2}, the course direction will alternate randomly.
\item \code{course_var} A means of adding minor variation to the \code{course} \code{data.frame} provided below;
will be ignored if \code{course} is missing or \code{NULL}. The argument \code{course_var} accepts a small
numeric (typically less than 0.005, for reasonable results), representing degrees latitude;
that number will be used to offset the \code{course} coordinates by \code{rnorm(1, mean=0, sd=course_var)}
in both the north-south and east-west directions.
\item \code{course} A \code{data.frame} of coordinates, representing the planned route for this
group of vessels. If provided, it must have two columns: \code{x} (longitude, decimal degrees
in which West is negative) and \code{y} (latitude, decimal degrees in which South is negative).
If not provided, the route will be randomly created as a linear line crossing the spatial grid.
}}

\item{toplot}{A Boolean; if \code{TRUE}, progress plots will be shown.}

\item{verbose}{A Boolean; if \code{TRUE}, updates will be printed to the Console.}
}
\value{
A \code{data.frame}, in which each row is a position fix for a transiting vessel,
with the following columns:
\itemize{
\item \code{vid} = A fake ID for the vessel
\item \code{type} = Vessel type
\item \code{speed} = Speed, in knots
\item \code{length} = Length
\item \code{width} = Beam width
\item \code{draft} = Draft
\item \code{datetime} = Datetime in UTC, with format \verb{yyyy-mm-dd hh:mm:ss}.
\item \code{x} = Longitude, decimal degrees (Western degrees negative).
\item \code{y} = Latitude, decimal degrees (Southern degrees negative).
\item \code{km} = The distance, in kilometers, between this and the next position fix.
}
}
\description{
Simulate a set of vessel transits on a spatial grid
}
