% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_grid.R
\name{summarize_grid}
\alias{summarize_grid}
\title{Summarize vessel grid by type, month, diel, and grid_id}
\usage{
summarize_grid(vgrid)
}
\arguments{
\item{vgrid}{desc}
}
\value{

}
\description{
Summarize vessel grid by type, month, diel, and grid_id
}
