% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tanker_route}
\alias{tanker_route}
\title{Tanker route}
\format{
An object of class \code{data.frame} with 82 rows and 2 columns.
}
\usage{
tanker_route
}
\description{
Tanker route
}
\keyword{datasets}
