% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tankers2030}
\alias{tankers2030}
\title{Interpolated version of simulated 2030 tanker traffic (July - October)
with grid cell associations.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 33389 rows and 16 columns.
}
\usage{
tankers2030
}
\description{
Interpolated version of simulated 2030 tanker traffic (July - October)
with grid cell associations.
}
\keyword{datasets}
