% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit_ellipses.R
\name{transit_ellipses}
\alias{transit_ellipses}
\title{Produce ellipses and test for close encounters}
\usage{
transit_ellipses(transits)
}
\arguments{
\item{transits}{A list of transit details produced during the \code{encounter_rate()} code.}
}
\value{
A \code{list}.
}
\description{
This is an internal \code{shipstrike} function not called directly by users.
}
