% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_grid.R
\name{vessel_grid}
\alias{vessel_grid}
\title{Interpolate vessel fixes and associate with grid cells}
\usage{
vessel_grid(grids, vessels, toplot = FALSE, verbose = TRUE)
}
\arguments{
\item{grids}{desc}

\item{vessels}{desc}

\item{toplot}{desc}

\item{verbose}{desc}
}
\value{

}
\description{
Interpolate vessel fixes and associate with grid cells
}
