% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collision_curve.R
\name{collision_curve}
\alias{collision_curve}
\title{Produce a P(Collision) ~ Speed curve}
\usage{
collision_curve(
  speeds = seq(0, 30, length = 1000),
  c1 = -0.2002002,
  c2 = 11.80241,
  asymptote = 0.9
)
}
\arguments{
\item{speeds}{desc}

\item{c1}{desc}

\item{c2}{desc}

\item{asymptote}{desc}
}
\value{

}
\description{
Produce a P(Collision) ~ Speed curve
}
