% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{Draw an ellipse}
\usage{
ellipse(x, y, width, height = width, theta = 0, npoints = 10, plot = FALSE)
}
\arguments{
\item{x}{Center x coordinate of ellipse.}

\item{y}{Center y coordinate of ellipse.}

\item{width}{Width of ellipse.}

\item{height}{Length of ellipse.}

\item{theta}{Angle/heading.}

\item{npoints}{Number of vertices to use to define the ellipse boundary.}

\item{plot}{Boolean; print a diagnostic plot?}
}
\value{
A list with three slots: \code{coords} (a \code{data.frame} of coordinates
defining ellipse boundary), \code{center} (the \code{x} and \code{y} inputs), and \code{theta} (the input angle).
}
\description{
Draw an ellipse
}
