% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_simulator.R
\name{encounter_simulator}
\alias{encounter_simulator}
\title{Encounter rate simulator}
\usage{
encounter_simulator(
  params.ship,
  v.whale,
  l.whale,
  w.whale,
  delta.sd,
  B = 100,
  encounter_meters = 0,
  speedy = TRUE,
  toplot = TRUE,
  save_records = FALSE,
  verbose = FALSE,
  plot_timeseries = FALSE
)
}
\arguments{
\item{params.ship}{desc}

\item{v.whale}{desc}

\item{l.whale}{desc}

\item{delta.sd}{desc}

\item{B}{desc}

\item{encounter_meters}{desc}

\item{speedy}{desc}

\item{toplot}{desc}

\item{save_records}{desc}

\item{verbose}{desc}

\item{plot_timeseries}{desc}
}
\value{

}
\description{
Encounter rate simulator
}
