% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encounter_tally.R
\name{encounter_tally}
\alias{encounter_tally}
\title{Tally encounters}
\usage{
encounter_tally(encounters)
}
\arguments{
\item{encounters}{Result of \code{encounter_simulator()}}
}
\value{

}
\description{
This includes processing near misses
}
