% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_outcomes.R
\name{gather_outcomes}
\alias{gather_outcomes}
\title{Gather and concatentate outcome \code{RData} files}
\usage{
gather_outcomes(outcome_dir, verbose = TRUE)
}
\arguments{
\item{outcome_dir}{Path (relative or absolute) to directory containing outcome files.}

\item{verbose}{Boolean; print updates to Console?}
}
\value{
A \code{data.frame}
}
\description{
This is not typically called by the user and currently deprecated.
}
