% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_seafloor.R
\name{get_seafloor}
\alias{get_seafloor}
\title{Get seafloor features at a spatial point}
\usage{
get_seafloor(lon, lat, seafloor, lat_range = 0.009009)
}
\arguments{
\item{lon}{The longitude (decimal degrees; West coordinates are negative) of interest.
This can be a single value or a vector of values.}

\item{lat}{The latitude (decimal degrees; South coordinates are negative).
This must be the same length as \code{lon}.}

\item{seafloor}{A \code{data.frame} with seafloor depths; required columns are
\code{x} (longitude, decimal degrees), \code{y} (latitude), and \code{layer} (with the seafloor depth).}

\item{lat_range}{The range around the provided coordinate, expressed as a numeric representing
degrees latitude, used to summarize seafloor features. The default represents 1 km north-south.}
}
\value{
A \code{dataframe}, with the number of rows equal to the length of \code{lon}, with the
following columns:
\itemize{
\item \code{z} Seafloor depth nearest to the coordinate.
\item \code{zmin} Minimum seafloor depth within \code{lat_range} of the coordinate.
\item \code{zmax} Maximum seafloor depth.
\item \code{zsd} Standard deviation of seafloor depth within \code{lat_range}.
}
}
\description{
Get seafloor features at a spatial point
}
