% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humpback_params.R
\name{humpback_params}
\alias{humpback_params}
\title{Humpback whale parameters to use in ship-strike analysis}
\usage{
humpback_params(
  length_min = 8.25,
  length_max = 15.5,
  length_mean = 11.85,
  length_sd = 1.32,
  width_factor = 0.33,
  speed_day_min = 0.13888,
  speed_day_max = 1.166,
  speed_day_mean = 0.7111,
  speed_day_sd = 0.20833,
  speed_night_min = 0.22035,
  speed_night_max = 1.85,
  speed_night_mean = 1.128,
  speed_night_sd = 0.2247,
  delta_day_mean = 0.8087532,
  delta_day_sd = 1.332026,
  delta_night_mean = 1.494988,
  delta_night_sd = 1.386166
)
}
\arguments{
\item{length_min}{Minimum whale length (meters) used to define a truncated normal distribution.}

\item{length_max}{Maximum whale length (meters) used to define a truncated normal distribution.}

\item{length_mean}{Mean whale length (meters).}

\item{length_sd}{Standard deviation in whale length (meters).}

\item{width_factor}{The scaling factor used to estimate whale fluke width based on length.
For exampl, a \code{width_factor} of 0.1 means a 20m whale has a fluke width of 2 meters.}

\item{speed_day_min}{Minimum daytime whale speed (meters/second)
used to define a truncated normal distribution.}

\item{speed_day_max}{Maximum daytime whale speed (meters/second)
used to define a truncated normal distribution.}

\item{speed_day_mean}{Mean daytime whale speed (meters/second)}

\item{speed_day_sd}{Standard deviation of daytime whale speed (meters/second)}

\item{speed_night_min}{Minimum nighttime whale speed (meters/second)
used to define a truncated normal distribution.}

\item{speed_night_max}{Maximum nighttime whale speed (meters/second)
used to define a truncated normal distribution.}

\item{speed_night_mean}{Mean nighttime whale speed (meters/second)}

\item{speed_night_sd}{Standard deviation of nighttime whale speed (meters/second)}

\item{delta_day_mean}{Mean of daytime whale track variability, defined as the
standard deviation of changes in whale course (log-transformed) every 60 seconds;
values will be drawn from a normal distribution with mean 0, then reverse-log-transformed.}

\item{delta_day_sd}{Standard deviation of track variability.}

\item{delta_night_mean}{Mean of nighttime whale track variability.}

\item{delta_night_sd}{Standard deviation of nighttime whale track variability.}
}
\value{
A named list of these inputs, without any changes at all. Pass this list to
\code{shipstrike::encounter_rate()} or \code{shipstrike::outcome_predict()}.
}
\description{
A list of parameters used for humpback whales in the analyses within \emph{Keen et al. (2023)}.
These parameters include all the details needed to estimate close-encounter rates and
predictions of whale-vessel interaction rates and collisions. You can use these same values
to get started in setting up your analysis,
then use this function to modify the values to fit your specific study.
}
\details{
The sources and rationale for these values are explaiend in \emph{Keen et al. (2023)}.
}
