% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lethality_curve.R
\name{lethality_curve}
\alias{lethality_curve}
\title{Produce a P(Lethality) ~ Speed curve}
\usage{
lethality_curve(
  speeds = seq(0, 30, length = 1000),
  c1 = -1.241241,
  c2 = 0.2712432,
  asymptote = 1
)
}
\arguments{
\item{speeds}{desc}

\item{c1}{desc}

\item{c2}{desc}

\item{asymptote}{desc}
}
\value{

}
\description{
Produce a P(Lethality) ~ Speed curve
}
