% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitigate_loop.R
\name{mitigate_loop}
\alias{mitigate_loop}
\title{Loop through months to find most efficacious mitigation window}
\usage{
mitigate_loop(
  outcomes,
  mitigation_duration = 1,
  reschedule = FALSE,
  vessels = c("Tanker > 100m", "Cargo > 100m", "Passenger > 100m",
    "LNG Canada tanker in-heel", "LNG Canada tanker in-product",
    "LNG Canada tug in-heel", "LNG Canada tug in-product", "Cedar LNG tanker in-heel",
    "Cedar LNG tanker in-product", "Cedar LNG tug in-heel", "Cedar LNG tug in-product")
)
}
\arguments{
\item{outcomes}{desc}

\item{mitigation_duration}{desc}

\item{reschedule}{desc}

\item{vessels}{desc}
}
\value{
desc
}
\description{
Loop through months to find most efficacious mitigation window
}
