% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_histograms.R
\name{outcome_histograms}
\alias{outcome_histograms}
\title{Create histograms summarizing whale-ship outcomes}
\usage{
outcome_histograms(
  outcomes,
  surface_var = "surface2",
  collision_var = "collision2.2",
  mortality_var = "mortality2.2",
  plot_all = FALSE,
  orientation = "vertical"
)
}
\arguments{
\item{outcomes}{A \code{data.frame} of results from the \code{shipstrike::gather_outcomes()} function.}

\item{surface_var}{The 'strike zone event' variable to use (\code{"surface1"} or \code{"surface2"})}

\item{collision_var}{The \code{collision} variable to use. See output of \code{outcome_predict()}.}

\item{mortality_var}{The \code{mortality} variable to use.  See output of \code{outcome_predict()}.}

\item{orientation}{Character; \code{"vertical"} or \code{"horizontal"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
This function produces pretty histograms of the posterior distributions of
outcome predictions (the default is to only show the outcomes for scenarios
in which strike zone is 1.5x draft and avoidance is a function of vessel speed).
}
