% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_histograms.R
\name{outcome_histograms}
\alias{outcome_histograms}
\title{Create histograms summarizing whale-ship outcomes}
\usage{
outcome_histograms(results)
}
\arguments{
\item{results}{A \code{data.frame} of results from the \code{shipstrike::outcome()} function.}
}
\value{

}
\description{
Create histograms summarizing whale-ship outcomes
}
