% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_map.R
\name{outcome_map}
\alias{outcome_map}
\title{Plot map of outcomes}
\usage{
outcome_map(
  outcomes_grid,
  base_map = NULL,
  varplot = "cooccurrence",
  varfacet = NULL,
  spp = NULL,
  years = NULL,
  channels = NULL,
  vessels = NULL,
  months = NULL,
  diels = NULL,
  col_min = "white",
  col_max = "darkorange4"
)
}
\arguments{
\item{outcomes_grid}{desc}

\item{base_map}{desc}

\item{varplot}{desc}

\item{varfacet}{desc}

\item{spp}{desc}

\item{years}{desc}

\item{channels}{desc}

\item{vessels}{desc}

\item{months}{desc}

\item{diels}{desc}

\item{col_min}{desc}

\item{col_max}{desc}
}
\value{

}
\description{
Plot map of outcomes
}
