% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_melt.R
\name{outcome_melt}
\alias{outcome_melt}
\title{Outcome melt}
\usage{
outcome_melt(
  mr,
  events = c("cooccurrence", "encounter", "surface", "surface2", "collision1.1",
    "collision1.2", "collision1.3", "collision1.4", "collision2.1", "collision2.2",
    "collision2.3", "collision2.4", "mortality1.1", "mortality1.2", "mortality1.3",
    "mortality1.4", "mortality2.1", "mortality2.2", "mortality2.3", "mortality2.4")
)
}
\arguments{
\item{mr}{desc}

\item{events}{desc}
}
\value{
desc
}
\description{
Outcome melt
}
