% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_shares.R
\name{outcome_shares}
\alias{outcome_shares}
\title{Share of risk by vessel type, channel, month, and diel period}
\usage{
outcome_shares(outcomes, verbose = TRUE)
}
\arguments{
\item{outcomes}{A \code{data.frame} of results from the \code{shipstrike::gather_outcomes()} function.}
}
\value{
A named \code{list}.
}
\description{
This function uses a bootstrapping routine (described in Keen et al. 2023)
to estimate the share of risk attributable to each level for vessel type,
channel, month, and diel period. It does this for each outcome type
(cooccurrence, close encounter, …, mortality).
}
