% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_table.R
\name{outcome_table}
\alias{outcome_table}
\title{Tabular summary of ship strike outcomes}
\usage{
outcome_table(outcomes)
}
\arguments{
\item{outcomes}{A \code{data.frame} of results from the \code{shipstrike::gather_outcomes()} function.}
}
\value{
A \code{data.frame}. See the output of \code{outcome_predict()}.
}
\description{
This function summarizes the predictions for each stage
of whale-vessel interaction, from cooccurrence to mortality.
}
