% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_table.R
\name{outcome_table}
\alias{outcome_table}
\title{Tabular summary of ship strike outcomes}
\usage{
outcome_table(results)
}
\arguments{
\item{results}{A \code{data.frame} of results from the \code{shipstrike::outcome()} function.}
}
\value{

}
\description{
Tabular summary of ship strike outcomes
}
