% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome_validation.R
\name{outcome_validation}
\alias{outcome_validation}
\title{Outcome validation simulation}
\usage{
outcome_validation(
  outcomes,
  eventcol = "mortality2.2",
  observations = 0,
  scaling_factor = 1,
  years = 10,
  months = 6:9,
  iterations = 1000,
  toplot = FALSE
)
}
\arguments{
\item{outcomes}{desc}

\item{observations}{desc}

\item{scaling_factor}{desc}

\item{years}{desc}

\item{months}{desc}

\item{iterations}{desc}

\item{toplot}{desc}
}
\value{

}
\description{
Outcome validation simulation
}
