% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_vessel.R
\name{simulate_vessel}
\alias{simulate_vessel}
\title{Simulate a set of vessels}
\usage{
simulate_vessel(grids, vessels, toplot = TRUE, verbose = TRUE)
}
\arguments{
\item{grids}{desc}

\item{vessels}{a list of lists}

\item{toplot}{desc}

\item{verbose}{desc}
}
\value{

}
\description{
Simulate a set of vessels
}
