% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_whale.R
\name{simulate_whale}
\alias{simulate_whale}
\title{Simulate whale density (mean / CV)}
\usage{
simulate_whale(grids, d_mean = 0.001, d_cv = 0.3)
}
\arguments{
\item{grids}{desc}

\item{d_mean}{desc}

\item{d_cv}{desc}
}
\value{

}
\description{
Simulate whale density (mean / CV)
}
