% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_grid.R
\name{summarize_grid}
\alias{summarize_grid}
\title{Summarize vessel grid by type, month, diel, and grid_id}
\usage{
summarize_grid(vgrid)
}
\arguments{
\item{vgrid}{An interpolated record of vessel position fixes,
as produced by \code{shipstrike::vessel_grid()}.}
}
\value{
A \code{list} with two slots:
\itemize{
\item \code{grid} holds a \code{data.frame} in which each row corresponds to a single grid cell
within a single month within a single diel period,
and the columns \code{speed}, \code{length}, \code{width}, and \code{draft} record the metrics for each vessel
that enters the cell, with the metric for each transit separated by an underscore (\verb{_}).
The column \code{n} indicates the number of transits that intersected the grid cell.

\item \code{params.ship} holds a \code{data.frame} in which each row corresponds to
vessel parameters during the transit of a single grid cell; there is a row for each
grid cell intersected across all transits, such that this \code{data.frame} represents
the spatially weighted distribution of characteristics to pass to the \code{encounter_simulator()}.
This dataset has three columns: \code{v.ship} (speed, in knots), \code{l.ship} (length, in meters), and \code{w.ship} (beam width, in meters).
Pass this slot to the \code{params.ship} input in \code{encounter_simulator()}.
}
}
\description{
This function is not typically called by the user.
This is essentially a wrapper for a \verb{dplyr::group_by() ... summarize()} sequence.
}
