% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transit_ellipses.R
\name{transit_ellipses}
\alias{transit_ellipses}
\title{Produce ellipses and test for close encounters}
\usage{
transit_ellipses(transits)
}
\arguments{
\item{transits}{desc}
}
\value{
desc
}
\description{
Produce ellipses and test for close encounters
}
