% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vessel_sun_angle.R
\name{vessel_sun_angle}
\alias{vessel_sun_angle}
\title{Calculate sun angle for a vessel based on \code{datetime}, \code{x}, and \code{y}}
\usage{
vessel_sun_angle(vessels)
}
\arguments{
\item{vessels}{desc}
}
\value{

}
\description{
Calculate sun angle for a vessel based on \code{datetime}, \code{x}, and \code{y}
}
